Fixed NSFW and Don't Overwrite Initially
allow_nsfw handling was broken because jsons was using too simple a
parser to handle booleans, and a structure design error in the
configuration objects needed to be corrected before a more complex
loader could be used.

Additionally, to avoid excessive duplication of files, the
first pass of the script
is now done without doing renaming of existing files. This isn't
exactly correct, but wil serve for now as a patch until a long term
stateless deduplication strategy can be implemented.